支持vue项目

+++ index.js
@@ -6,6 +6,9 @@ const path = require('path');
 const codeCheckDir = __dirname;
 
 async function copyFile(framework, filename) {
+  if (fs.existsSync(filename)) {
+    return;
+  }
   await fs.copy(path.resolve(codeCheckDir, `${framework}/${filename}`), filename);
 }
 
@@ -23,8 +26,8 @@ async function task(framework) {
   await execa.shell(`npm i -D prettier pretty-quick eslint-config-prettier eslint-plugin-prettier \
   eslint-config-standard eslint-plugin-standard eslint-plugin-import eslint-plugin-node eslint-plugin-promise \
   husky  lint-staged `, {
-      stdio: 'inherit',
-    });
+    stdio: 'inherit',
+  });
 
   // 拷贝配置文件
   await copyFile(framework, '.eslintrc.js');
@@ -53,7 +56,43 @@ async function task(framework) {
   )
 }
 
+async function vueInit() {
+  const framework = 'vue';
+  // 安装依赖
+  await execa.shell(`npm i -D prettier pretty-quick eslint-config-prettier eslint-plugin-prettier husky lint-staged`, {
+    stdio: 'inherit',
+  });
+
+  // 拷贝配置文件
+  await copyFile(framework, '.eslintignore');
+  await copyFile(framework, '.prettierrc.js');
+  await copyFile(framework, '.prettierignore');
+
+  //修改package.json文件
+  await packageJson(
+    'husky',
+    {
+      "hooks": {
+        "pre-commit": "lint-staged"
+      }
+    }
+  );
+  await packageJson(
+    "lint-staged",
+    {
+      "*.{js,vue}": [
+        "eslint --fix",
+        "git add"
+      ],
+      "*": "pretty-quick --staged"
+    }
+  )
+
+}
+
 const args = process.argv.slice(2);
-if(args[0] === '--init'){
+if (args[0] === 'react') {
   task('react');
+} else if (args[0] === 'vue') {
+  vueInit();
 }
